streaming/requirements.txt
pulsar-client==2.10.1
openai
pandas 
numpy 
flask 
flask_cors 
cassandra-driver 
tiktoken
tenacity